+++ docs/blog/posts/release0-38-0.md
@@ -14,7 +14,7 @@ This is the second big feature release this month after last week's [command pal
 
 <!-- more -->
 
-The [TextArea](../../widgets/text_area.md) has finally landed.
+The [TextArea](../../widgets/text_editor.md) has finally landed.
 I know a lot of folk have been waiting for this one.
 Textual's TextArea is a fully-featured widget for editing code, with syntax highlighting and line numbers.
 It is highly configurable, and looks great.
