+++ src/textual/widgets/__init__.py
@@ -41,7 +41,7 @@ if typing.TYPE_CHECKING:
     from ._switch import Switch
     from ._tabbed_content import TabbedContent, TabPane
     from ._tabs import Tab, Tabs
-    from ._text_area import TextArea
+    from ._text_editor import TextEditor
     from ._tooltip import Tooltip
     from ._tree import Tree
     from ._welcome import Welcome
@@ -80,7 +80,7 @@ __all__ = [
     "TabbedContent",
     "TabPane",
     "Tabs",
-    "TextArea",
+    "TextEditor",
     "RichLog",
     "Tooltip",
     "Tree",
