+++ src/textual/widgets/__init__.pyi
@@ -33,7 +33,7 @@ from ._tabbed_content import TabbedContent as TabbedContent
 from ._tabbed_content import TabPane as TabPane
 from ._tabs import Tab as Tab
 from ._tabs import Tabs as Tabs
-from ._text_area import TextArea as TextArea
+from ._text_editor import TextEditor as TextEditor
 from ._tooltip import Tooltip as Tooltip
 from ._tree import Tree as Tree
 from ._welcome import Welcome as Welcome
