--- src/textual/widgets/text_area.py
@@ -1,37 +1,0 @@
-from textual._text_area_theme import TextAreaTheme
-from textual.document._document import (
-    Document,
-    DocumentBase,
-    EditResult,
-    Location,
-    Selection,
-)
-from textual.document._languages import BUILTIN_LANGUAGES
-from textual.document._syntax_aware_document import SyntaxAwareDocument
-from textual.widgets._text_area import (
-    Edit,
-    EndColumn,
-    Highlight,
-    HighlightName,
-    LanguageDoesNotExist,
-    StartColumn,
-    ThemeDoesNotExist,
-)
-
-__all__ = [
-    "BUILTIN_LANGUAGES",
-    "Document",
-    "DocumentBase",
-    "Edit",
-    "EditResult",
-    "EndColumn",
-    "Highlight",
-    "HighlightName",
-    "LanguageDoesNotExist",
-    "Location",
-    "Selection",
-    "StartColumn",
-    "SyntaxAwareDocument",
-    "TextAreaTheme",
-    "ThemeDoesNotExist",
-]
