+++ src/textual/widgets/text_editor.py
@@ -1,0 +1,37 @@
+from textual._text_editor_theme import TextEditorTheme
+from textual.document._document import (
+    Document,
+    DocumentBase,
+    EditResult,
+    Location,
+    Selection,
+)
+from textual.document._languages import BUILTIN_LANGUAGES
+from textual.document._syntax_aware_document import SyntaxAwareDocument
+from textual.widgets._text_editor import (
+    Edit,
+    EndColumn,
+    Highlight,
+    HighlightName,
+    LanguageDoesNotExist,
+    StartColumn,
+    ThemeDoesNotExist,
+)
+
+__all__ = [
+    "BUILTIN_LANGUAGES",
+    "Document",
+    "DocumentBase",
+    "Edit",
+    "EditResult",
+    "EndColumn",
+    "Highlight",
+    "HighlightName",
+    "LanguageDoesNotExist",
+    "Location",
+    "Selection",
+    "StartColumn",
+    "SyntaxAwareDocument",
+    "TextEditorTheme",
+    "ThemeDoesNotExist",
+]
