+++ tests/document/test_document.py
@@ -1,6 +1,6 @@
 import pytest
 
-from textual.widgets.text_area import Document
+from textual.widgets.text_editor import Document
 
 TEXT = """I must not fear.
 Fear is the mind-killer."""
