+++ tests/document/test_document_delete.py
@@ -1,6 +1,6 @@
 import pytest
 
-from textual.widgets.text_area import Document, EditResult
+from textual.widgets.text_editor import Document, EditResult
 
 TEXT = """I must not fear.
 Fear is the mind-killer.
