--- tests/snapshot_tests/snapshot_apps/text_area_unfocus.py
@@ -1,17 +1,0 @@
-"""Tests the rendering of the TextArea for all supported languages."""
-from textual.app import App, ComposeResult
-from textual.widgets import TextArea
-
-
-class TextAreaUnfocusSnapshot(App):
-    AUTO_FOCUS = None
-
-    def compose(self) -> ComposeResult:
-        text_area = TextArea()
-        text_area.cursor_blink = False
-        yield text_area
-
-
-app = TextAreaUnfocusSnapshot()
-if __name__ == "__main__":
-    app.run()
